week_06_07/project1_rag/requirements.txt
faiss-cpu
pdfplumber
PyPDF2
numpy
pandas
sentence-transformers
transformers
accelerate
torch
